removed exit() from errors
the hsm_ctx_t structure now contains three new attributes;
int error
const char *error_action and
const char *error_message

if an error is encountered (mostly PKCS#11 errors, later we could perhaps also use this for other errors such as memory), and the value of error is zero, it is set to the CK_RV return value of the failing pkcs11 call. The error_action will then be a static string containing information about what libhsm was trying to do, and the static string error_message what the actual error was

if no errors were encountered, these values are 0, NULL and NULL.

it is only set if not already non-zero so if we do multiple pkcs11 calls, the first error is probably the most important one.

all ctx arguments are now no longer const, since any error can modify them.

for applications: if you get an unexpected NULL value, or an unexpected 0, check you context for ctx->error. You could of course check it after any call that uses the ctx, which would be more prudent.


git-svn-id: svn+ssh://svn.opendnssec.org/svn/dnssec@1174 b8cecbbd-7941-0410-8e65-efea935c1ce0