SCA6000 complains if the CKA_PUBLIC_EXPONENT is not included. Should default to 65537. Maybe the SCA6000 is only v2.11
From PKCS#11 API:
Note: Implementations strictly compliant with version 2.11 or prior versions may generate an error if this attribute is omitted from the template. 
Experience has shown that many implementations of 2.11 and prior did allow the CKA_PUBLIC_EXPONENT attribute to be omitted from 
the template, and behaved as described above.


git-svn-id: svn+ssh://svn.opendnssec.org/svn/dnssec@3963 b8cecbbd-7941-0410-8e65-efea935c1ce0