OPENDNSSEC-260: Signer crashes when hsm_create_context() failed to create context
This will make the signer deal a bit much better with a failing
hsm_create_context(). The idea is that if it fails, the signer engine will be
reloaded. This will reopen the connection to the HSM, stop all workers,
including drudgers, and restart them again.

If it is failing everytime, this will make the signer do a 100% job on
reloading and retrying. That is because on a reload, the zones are updated,
and on a zone update, the signconf is reloaded. Thus, task is rescheduled to
now, fails, signer reloads, task is rescheduled, fails, signer reloads, etc...

So, engine_update_zones should only reschedule task if zonelist is updated.
Hence, the added zl_changed parameter for the engine_update_zones() function,
to honor the task backoff timers.

This influences the issue OPENDNSSEC-228 "Make ods-signer update reload
signconf even if zonelist has not change". So in cmdhan



git-svn-id: svn+ssh://svn.opendnssec.org/svn/dnssec@6450 b8cecbbd-7941-0410-8e65-efea935c1ce0